refactor(client): extract the Query tab into its own OnPush component (A17.9b-6a)
First of the six tabs to leave the shell, and deliberately the simplest: the Query
tab is read-only (advanced MongoDB-style query + semantic recall, no create/edit
forms), so it proves the "tab -> component over a spaceId input" pattern before the
heavier record tabs.

- query-tab.component.ts (QueryTabComponent): OnPush, standalone. Owns the query/recall
  forms + results + the run/clear/format methods, moved verbatim. Talks only to BrainApi
  (queryBrain/recallBrain) and BrainStore (the recall "filter by type" options). The
  active space id is a required signal input, read at call time so a mid-flight space
  switch cannot stale an in-flight request (activeSpaceId() -> spaceId()). The query-only
  CSS moved into the component's styles.
- Shell renders `@if (activeTab() === 'query') { <app-query-tab [spaceId]=... /> }` so the
  inactive tab is not instantiated. Moved state/methods/CSS removed; now-unused
  Query*/Recall* type imports dropped.

5 tests pin the OnPush flag, formatQueryDoc, the panel render, and the two guard paths
(blank recall query is a no-op; invalid-JSON filter -> form error, no API call).

brain.component.ts 2845 -> 2431. Client suite 169 -> 174. tsc --noUnusedLocals clean on
the brain files; prod AOT build warning-free bar the pre-existing qrcode CJS bailout.
